+++ compiler/src/symbol_table.c
@@ -30,7 +30,11 @@ int remove_local_symbol_table(SymbolTables* symtables) {
     }
     free(table_keys);
     free_hashmap(table);
-    symtables->local_tables = (HashMap**) (realloc(symtables->local_tables, sizeof(HashMap*) * symtables->no_local_tables)) ;
+    if (symtables->no_local_tables == 0) {
+        free(symtables->local_tables);
+    } else {
+        symtables->local_tables = (HashMap**) (realloc(symtables->local_tables, sizeof(HashMap*) * symtables->no_local_tables)) ;
+    }
     return 0;
 }
 
@@ -77,37 +81,41 @@ int freeAll(SymbolTables* symt) {
 
 /*************STEntry********* */
 
-STEntry* create_stentry(STETYPE dtype) {
+STEntry* create_stentry(STETYPE dtype, char* name) {
     STEntry* ste = (STEntry*) calloc(1, sizeof(STEntry));
     ste->dtype = dtype;
+    ste->name = name;
 }
 
 int add_array_layer(STEntry* ste, int len) {
     if (len <= 0) {
         return -1;
     }
-    if (ste->array_depth == 0) {
-        ste->total_array_length = 1;
+    if (ste->is_array == 0 ) {
+        // adding first layer
+        ste->is_array = 1;
+        ste->value.arrval = (DARRAY*) calloc(1, sizeof(DARRAY));
+        ste->value.arrval->arr_max_pos = 1;
     }
-    ste->total_array_length *= len;
-    ste->array_depth ++;
-    ste->array_length = (int*) realloc(ste->array_length, sizeof(int) * ste->array_depth);
-    ste->array_length[ste->array_depth - 1] = len;
+    ste->value.arrval->arr_max_pos *= len;
+    ste->value.arrval->arr_depth ++;
+    ste->value.arrval->arr_lengths = (int*) realloc(ste->value.arrval->arr_lengths, sizeof(int) * ste->value.arrval->arr_depth);
+    ste->value.arrval->arr_lengths[ste->value.arrval->arr_depth - 1] = len;
     switch (ste->dtype)
     {
     case DOUBLE:
-        free(ste->arr);
-        ste->arr = (char*) (malloc(sizeof(double) * ste->total_array_length));
+        free(ste->value.arrval->arr);
+        ste->value.arrval->arr = (char*) (malloc(sizeof(double) * ste->value.arrval->arr_max_pos));
         break;
     
     case INT:
-        free(ste->arr);
-        ste->arr = (char*) (malloc(sizeof(lli) * ste->total_array_length));
+        free(ste->value.arrval->arr);
+        ste->value.arrval->arr = (char*) (malloc(sizeof(lli) * ste->value.arrval->arr_max_pos));
         break;
     
     case BOOL:
-        free(ste->arr);
-        ste->arr = (char*) (malloc(sizeof(short) * ste->total_array_length));
+        free(ste->value.arrval->arr);
+        ste->value.arrval->arr = (char*) (malloc(sizeof(short) * ste->value.arrval->arr_max_pos));
         break;
 
     default:
@@ -117,118 +125,162 @@ int add_array_layer(STEntry* ste, int len) {
 }
 
 lli update_double(STEntry* ste, double val,int pos) {
-    if (pos > -1 && ste->array_depth == 0) { // not array but still a position is given
+    if (pos > -1 && ste->is_array == 0) { // not array but still a position is given
         return LLONG_MIN + 1;
     } 
-    if (pos == -1 && ste->array_depth != 0) { // array but no position is given
+    if (pos == -1 && ste->is_array != 0) { // array but no position is given
         return LLONG_MIN;
     } 
     if (ste->dtype != DOUBLE) { // wrong type
         return LLONG_MIN + 2;
     }
-    if (pos >= ste->total_array_length) {
+    if (pos >= ste->value.arrval->arr_max_pos) {
         return LLONG_MIN + 3; // out of bounds
     }
     if (pos == -1) {
         ste->value.dval = val;
     } else {
-        *((double*)(ste->arr) + pos) = val;
+        *((double*)(ste->value.arrval->arr) + pos) = val;
     }
     return 0;
 }
 
 lli update_int(STEntry* ste, lli val,int pos) {
-    if (pos > -1 && ste->array_depth == 0) { // not array but still a position is given
+    if (pos > -1 && ste->is_array == 0) { // not array but still a position is given
         return LLONG_MIN + 1;
     } 
-    if (pos == -1 && ste->array_depth != 0) { // array but no position is given
+    if (pos == -1 && ste->is_array != 0) { // array but no position is given
         return LLONG_MIN;
     } 
     if (ste->dtype != INT) { // wrong type
         return LLONG_MIN + 2;
     }
+    if (pos >= ste->value.arrval->arr_max_pos) {
+        return LLONG_MIN + 3; // out of bounds
+    }
     if (pos == -1) {
         ste->value.lval = val;
     } else {
-        *((lli*)(ste->arr) + pos) = val;
+        *((lli*)(ste->value.arrval->arr) + pos) = val;
     }
     return 0;
 }
 
 lli update_bool(STEntry* ste, short val,int pos) {
-    if (pos > -1 && ste->array_depth == 0) { // not array but still a position is given
+    if (pos > -1 && ste->is_array == 0) { // not array but still a position is given
         return LLONG_MIN + 1;
     } 
-    if (pos == -1 && ste->array_depth != 0) { // array but no position is given
+    if (pos == -1 && ste->is_array != 0) { // array but no position is given
         return LLONG_MIN;
     } 
     if (ste->dtype != BOOL) { // wrong type
         return LLONG_MIN + 2;
     }
+    if (pos >= ste->value.arrval->arr_max_pos) {
+        return LLONG_MIN + 3; // out of bounds
+    }
     if (pos == -1) {
         ste->value.lval =val != 0;
     } else {
-        *((short*)(ste->arr) + pos) = val;
+        *((short*)(ste->value.arrval->arr) + pos) = val;
     }
     return 0;
 }
 
 lli* get_double(STEntry* ste, int pos) {
-    if (pos > -1 && ste->array_depth == 0) { // not array but still a position is given
+    if (pos > -1 && ste->is_array == 0) { // not array but still a position is given
         return (lli*)(INT_MIN + 1);
     } 
-    if (pos == -1 && ste->array_depth != 0) { // array but no position is given
-        return (lli*)ste->arr;
+    if (pos == -1 && ste->is_array != 0) { // array but no position is given
+        return (lli*)ste->value.arrval->arr;
     } 
     if (ste->dtype != DOUBLE) { // wrong type
         return (lli*)(INT_MIN + 2);
     }
+    if (pos >= ste->value.arrval->arr_max_pos) {
+        return (lli*)LLONG_MIN + 3; // out of bounds
+    }
     if (pos == -1) {
         return (lli*)&(ste->value.dval);
     } else {
-        return (lli*)((double*)ste->arr + pos);
+        return (lli*)((double*)ste->value.arrval->arr + pos);
     }
     return 0;
 }
 
 lli* get_int(STEntry* ste, int pos) {
-    if (pos > -1 && ste->array_depth == 0) { // not array but still a position is given
+    if (pos > -1 && ste->is_array == 0) { // not array but still a position is given
         return (lli*)(INT_MIN + 1);
     } 
-    if (pos == -1 && ste->array_depth != 0) { // array but no position is given
-        return (lli*)ste->arr;
+    if (pos == -1 && ste->is_array != 0) { // array but no position is given
+        return (lli*)ste->value.arrval->arr;
     } 
     if (ste->dtype != INT) { // wrong type
         return (lli*)(INT_MIN + 2);
     }
+    if (pos >= ste->value.arrval->arr_max_pos) {
+        return (lli*)LLONG_MIN + 3; // out of bounds
+    }
     if (pos == -1) {
         return &(ste->value.lval);
     } else {
-        return ((lli*)ste->arr + pos);
+        return ((lli*)ste->value.arrval->arr + pos);
     }
     return 0;
 }
 
 lli* get_bool(STEntry* ste, int pos) {
-    if (pos > -1 && ste->array_depth == 0) { // not array but still a position is given
+    if (pos > -1 && ste->is_array == 0) { // not array but still a position is given
         return (lli*)(INT_MIN + 1);
     } 
-    if (pos == -1 && ste->array_depth != 0) { // array but no position is given
-        return (lli*)ste->arr;
+    if (pos == -1 && ste->is_array != 0) { // array but no position is given
+        return (lli*)ste->value.arrval->arr;
     } 
     if (ste->dtype != BOOL) { // wrong type
         return (lli*)(INT_MIN + 2);
     }
+    if (pos >= ste->value.arrval->arr_max_pos) {
+        return (lli*)LLONG_MIN + 3; // out of bounds
+    }
     if (pos == -1) {
         return &(ste->value.lval);
     } else {
-        return (lli*)((short*)(ste->arr) + pos);
+        return (lli*)((short*)(ste->value.arrval->arr) + pos);
     }
     return 0;
 }
 
 void free_ste(STEntry* ste) {
-    if (ste->arr != NULL) free(ste->arr);
-    if (ste->array_length != NULL) free(ste->array_length);
+    if (ste->is_array != 0) {
+        if (ste->value.arrval->arr != NULL) free(ste->value.arrval->arr);
+        if (ste->value.arrval->arr_lengths != NULL) free(ste->value.arrval->arr_lengths);
+        free(ste->value.arrval);
+    }
     free(ste);
+}
+
+int sprintf_ste(STEntry* ste, char* val, int len) {
+    int start = 0;
+    int total_len = len;
+    len = total_len - start;
+    switch (ste->dtype)
+    {
+        case DOUBLE:
+        start += snprintf(val + start, len, "(double)");
+        break;
+        case INT:
+        start += snprintf(val + start, len, "(int)");
+        break;
+        case BOOL:
+        start += snprintf(val + start, len, "(bool)");
+        break;
+        default:
+        break;
+    }
+    len = total_len - start;
+    if (ste->is_array) {
+        start += snprintf(val + start, len, "[]");
+    }
+    start += snprintf(val + start, len, "(%s)", ste->name);
+    return start;
 }